+++ Runtime/Authoring/InspectorReadOnly.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace DOTSAnimation.Authoring
-{
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Field | AttributeTargets.Property)]
-    public class InspectorReadOnly : Attribute{}
-}